+++ base/src/systems/username_invalid_move_user_to_needsusername.rs
@@ -1,9 +1,7 @@
 use shared::prelude::*;
 
-pub fn username_invalid(
+pub fn username_invalid_move_user_to_needsusername(
     mut username_invalid_rx: EventReader<UsernameInvalidEvent>,
-    mut text_event_tx: EventWriter<TextEvent>,
-    mut show_prompt_rx: EventWriter<ShowPromptEvent>,
     mut query: Query<&mut UserSessionData>,
 ) {
     for ev in username_invalid_rx.read() {
@@ -15,9 +13,5 @@ pub fn username_invalid(
             }
         };
         user_session_data.status = UserStatus::NeedUsername;
-
-        text_event_tx.send(TextEvent::from_str(ev.0, 
-            "{{196}}Sorry, that's an invalid username. {{7}}Usernames must be:\n * Between 3 and 16 characters long\n * Contain only letters, numbers, and underscores\n * Starts with a letter"));
-        show_prompt_rx.send(ShowPromptEvent(ev.0));
     }
 }
